requirements.txt
BeautifulSoup==3.2.1
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
django-adminlte2==0.1.6
django-appconf==1.0.2
django-crispy-forms==1.6.1
django-datatable==0.3.1
django-datetime-widget-fa==0.10.0
django-formsets-bootstrap3==0.2.2
django-mssql==1.8
django-pyodbc==1.1.0
django-pyodbc-azure==1.11.0.0
Django-Select2==5.11.1
docopt==0.6.2
idna==2.5
pipreqs==0.4.9
pyodbc==4.0.17
pytz==2017.2
requests==2.18.3
urllib3==1.22
yarg==0.1.9
